-Fixed error in rs_simple which resulted in an infinite loop when trying to retreive the config and the initial buffer was to small.
-Modified CMakeLists.txt to install binaries

-Made lio_* commands automatically detect LFS FUSE mounts by examing /proc/mounts  and translate paths
accordingly.  For exaample if LFS is mounted on /lio/lfs and the CWD is /lio/lfs/dummy then the user
can display the dir contents using:

1) ls
2) lio_ls /dummy
3) lio_ls @:/dummy
4) lio_ls .

For most commands you don't need to use the "@:" notation to signify a local vs LFS mount.  But it is
required for lio_cp.  The first 3 were available before.  Option 4 is new.  This makes moving between
lio_* commands and traditional linux commands trivial.