Fixed bug in calculation of child_last_page
The the child_last_page should coorespond to the offset of the last page in the child segment.  In one instance it
was being calculated using the file size and not the offset.

This exposed another bug resulting in blank stripes not being handle properly which was fixed as well.