Fixed nasty readdir bug resulting in missing or duplicated files.
This bug only pops up on large directories where the number of files doesn't
fit in the FUSE buffer resulting in the readdir() operation being called multiple times.
In this case the dentry offset was being caluclated incoreeectly leading to missing files.

The dentries are all stored on a stack and the normal use case is upon reentry from a buffer fill
the requested offset is the last item in the stack.  The old code started from the beginning of
the list and iterated to the end leading to an n^2 scaling.  This has been tweaked to move to the
end and back up to the requested offset.