Update statfs for OSX
The statvfs struct under OSX unconditionally uses 'unsigned int' for the
number of blocks. If the blocks are 4kB, this means the maximum
reportable size of the filesystem is 4kB * 2^32 = 8TB. If our filesystem
is larger than that, try and increase the block size so the total number
of blocks will fit within an unsigned int. The other side is that
there's a maximum block size of 128kB. If that happens, clamp the block
size and the number of blocks to their limits.

The sum of these limits is that the largest representable size of the
filesystem is 512TB (128kB * 2^32). Big Data Problems, I guess.